+++ BirdsEye/sim/sim_prototypes.h
@@ -160,6 +160,7 @@ bool sdEnsureTracksFolder();
 void sdPerformFormat();
 bool buildTrackList();
 int parseTrackFile(char* filepath);
+int parseTrackFileEntry(char* filepath, const char* trackKey);
 
 // ---- settings.ino ----
 bool createDefaultSettings();